# backend/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MicroEra Wiki MVP — Enterprise Data Pipeline API",
    "version": "0.1.0",
    "description": "企业数据源管道 API：Upload → Parse → Document → Chunk → Embed → Vector → Retrieve → LLM\n\n支持 PDF / Word / Markdown / TXT / PNG 等多格式导入，MarkItDown 统一解析，BGE-M3 向量化，Milvus 向量检索，Ollama LLM 问答。",
    "contact": {
      "name": "MicroEra Wiki Team"
    }
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local development server" }
  ],
...
  "paths": {
...
    "/api/pipeline/parse": {
      "post": {
        "tags": ["Pipeline"],
        "summary": "仅解析 — 将文件内容转为 Markdown",
        "description": "支持 PDF、DOCX、MD、TXT、PNG 等格式。MarkItDown 概念：所有格式统一转换为 Markdown。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": { "type": "string", "description": "原始文本内容" },
                  "fileName": { "type": "string", "description": "文件名（用于格式检测）" },
                  "filePath": { "type": "string", "description": "服务器上的文件路径" }
                }
              },
              "examples": {
                "markdown": {
                  "summary": "解析 Markdown 文本",
                  "value": {
                    "content": "# 标题\n\n这是一段测试文本。",
                    "fileName": "test.md"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "解析结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "markdown": { "type": "string", "description": "统一的 Markdown 内容" },
                    "fullLength": { "type": "integer" },
                    "metadata": { "type": "object" },
                    "propertyCount": { "type": "integer" },
                    "properties": { "type": "array" },
                    "warnings": { "type": "array", "items": { "type": "string" } }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/pipeline/chunk": {
      "post": {
        "tags": ["Pipeline"],
        "summary": "仅分块 — 将文本拆分为可嵌入的块",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": {
                  "text": { "type": "string", "description": "要分块的文本" },
                  "strategy": {
                    "type": "string",
                    "enum": ["fixed", "paragraph", "sentence", "markdown"],
                    "default": "markdown"
                  },
                  "chunkSize": { "type": "integer", "default": 512 },
                  "overlap": { "type": "integer", "default": 64 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "分块结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "chunkCount": { "type": "integer" },
                    "config": { "type": "object" },
                    "chunks": { "type": "array" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/pipeline/embed": {
      "post": {
        "tags": ["Pipeline"],
        "summary": "仅嵌入 — 生成向量并存入向量库",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": { "type": "string" },
                  "texts": { "type": "array", "items": { "type": "string" } },
                  "entryId": { "type": "integer", "description": "关联的条目 ID（用于向量存储）" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "嵌入结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "vectorCount": { "type": "integer" },
                    "dimension": { "type": "integer" },
                    "storedInVectorDb": { "type": "integer" },
                    "model": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/pipeline/import": {
      "post": {
        "tags": ["Pipeline"],
        "summary": "完整导入管道 — Upload → Parse → Metadata → Chunk → Embed → Vector",
        "description": "支持 multipart/form-data 文件上传或 JSON body（filePath/content）。\n\n管道步骤：\n1. 文件上传/接收\n2. 格式检测与 MarkItDown 解析\n3. 创建元数据条目\n4. 智能分块（markdown-aware）\n5. Ollama 嵌入向量生成\n6. Milvus/内存向量库存储",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary", "description": "上传文件" },
                  "metadata": { "type": "string", "description": "JSON 格式的条目元数据" },
                  "skipEmbedding": { "type": "string", "enum": ["true", "false"] },
                  "chunkConfig": { "type": "string", "description": "JSON 格式的分块配置" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "filePath": { "type": "string" },
                  "content": { "type": "string" },
                  "fileName": { "type": "string" },
                  "metadata": { "type": "object" },
                  "skipEmbedding": { "type": "boolean" },
                  "chunkConfig": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "导入成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "entryId": { "type": "integer" },
                    "parse": { "type": "object" },
                    "chunks": { "type": "object" },
                    "embedding": { "type": "object" },
                    "vector": { "type": "object" },
                    "timing": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/pipeline/import/string": {
      "post": {
        "tags": ["Pipeline"],
        "summary": "从字符串内容导入",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content"],
                "properties": {
                  "content": { "type": "string" },
                  "fileName": { "type": "string" },
                  "metadata": { "type": "object" },
                  "skipEmbedding": { "type": "boolean" },
                  "chunkConfig": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "导入结果"
          }
        }
      }
    },
    "/api/pipeline/import/batch": {
      "post": {
        "tags": ["Pipeline"],
        "summary": "批量导入 — 扫描目录并导入所有支持的文件",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["dirPath"],
                "properties": {
                  "dirPath": { "type": "string", "description": "要扫描的目录路径" },
                  "pattern": { "type": "string", "default": "*", "description": "文件名匹配模式" },
                  "skipEmbedding": { "type": "boolean" },
                  "chunkConfig": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "批量导入结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": { "type": "integer" },
                    "succeeded": { "type": "integer" },
                    "failed": { "type": "integer" },
                    "results": { "type": "array" },
                    "totalTimeMs": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/pipeline/search": {
      "post": {
        "tags": ["Pipeline"],
        "summary": "语义搜索 — Embedding → TopK → Context",
        "description": "将查询文本向量化，在向量库中检索最相关的条目。自动降级：Milvus → 内存向量 → 关键词搜索。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": { "type": "string", "description": "自然语言查询" },
                  "topK": { "type": "integer", "default": 10 },
                  "isInternal": { "type": "boolean", "default": false }
                }
              },
              "examples": {
                "semantic": {
                  "summary": "语义搜索材料性质",
                  "value": { "query": "MOF 材料的可及表面积和孔径", "topK": 5 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "搜索结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "query": { "type": "string" },
                    "totalResults": { "type": "integer" },
                    "queryTimeMs": { "type": "integer" },
                    "source": { "type": "string", "enum": ["milvus", "memory_vector", "keyword"] },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "entryId": { "type": "integer" },
                          "title": { "type": "string" },
                          "score": { "type": "number" },
                          "summary": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}